+++ src/main/java/com/mo/economy_system/core/playerattributes_system/infection/PlayerInfectionClientSync.java
@@ -6,7 +6,7 @@ import net.minecraft.server.level.ServerPlayer;
 import net.minecraftforge.network.NetworkDirection;
 
 public class PlayerInfectionClientSync {
-    public static void sendInfectionDataToClient(ServerPlayer player, int currentInfection) {
+    public static void sendInfectionDataToClient(ServerPlayer player, float currentInfection) {
         Packet_SyncInfectionData packet = new Packet_SyncInfectionData(currentInfection);
         EconomySystem_NetworkManager.INSTANCE.sendTo(packet, player.connection.connection, NetworkDirection.PLAY_TO_CLIENT);
     }
